requirements.txt
# Core framework dependencies
langgraph>=0.2.40
langchain-core>=0.3.15

# LLM provider clients
langchain-google-genai>=2.0.4
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
google-generativeai>=0.8.0

# Utilities and infrastructure
tenacity>=9.0.0
python-dotenv>=1.0.1
httpx>=0.27.2

# Git and GitHub integration
GitPython>=3.1.43
pygithub>=2.4.0

# CLI and UI
typer>=0.12.5
rich>=13.8.0

# Data processing and validation
jsonschema>=4.23.0
datasets>=3.0.0

# Code analysis and highlighting
pygments>=2.18.0

# YAML processing
ruamel.yaml>=0.18.6

